add silhouette notebook

+++ lib/render/camera.py
@@ -26,7 +26,7 @@ class Camera:
         mat = get_translation(self.dist)
         mat = mat @ get_rotation_x(self.elev)
         mat = mat @ get_rotation_y(self.azim)
-        return mat
+        return mat.astype(np.float32)
 
     def get_camera_to_world(self):
         return np.linalg.inv(self.get_world_to_camera()).astype(np.float32)
